requirements.txt
asgiref==3.2.10
coverage==5.3
Django==3.1.1
django-cleanup==5.0.0
django-nose==1.4.7
djangorestframework==3.11.1
nose==1.3.7
psycopg2-binary==2.8.6
pytz==2020.1
sqlparse==0.3.1
